OneDrive/Desktop/PlantDisease/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
scikit-learn==1.4.0
python-dotenv==1.0.0
gunicorn==21.2.0 